SoraRTCStats を MessageConverter に移動する
